// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/Proceso.ts
export interface Proceso {
  tipoMaquina: string;
  claveProceso: string;
  descripcion: string;
  estatus: boolean;
  tratamiento: boolean;
  subProceso: string;
  tintas: string;
  conSuaje: boolean;
}
export class Proceso {
  tipoMaquina: string;
  claveProceso: string;
  descripcion: string;
  estatus: boolean;
  tratamiento: boolean;
  subProceso: string;
  tintas: string;
  conSuaje: boolean;
}